requirements-test.txt
# Testing dependencies
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-html>=3.2.0
factory-boy>=3.3.0
faker>=19.6.2
coverage>=7.3.0
django-coverage-plugin>=3.1.0
django-test-plus>=2.2.0
model-bakery>=1.12.0

# Accessibility testing
playwright>=1.40.0
pytest-playwright>=0.4.3
axe-playwright-python>=0.1.3

# Visual regression testing
pillow>=10.0.0
pytest-image-diff>=0.0.14